wip: #70 scrollable parent reordering
This tries to demo whether dragging works correctly when the parent is
scrolled to a different position. Note that since we don't want the
parent to capture the pan gesture, we actually disable the ScrollView
from capturing, and manually scroll via a `useEffect` hook.

It doesn't seem #70 reproduces in this example. When dragging,
everything looks correct, even the first time.