fix: #76 Remove flicker on reorder (#80)
There was a timing issue between when `pan.setValue(0)` was called
relative to when the newly-reordered list is rendered. I investigated a
bunch of ways to fix this, including mirroring the list, but in the end
the best way is to hide the dragged object briefly when the list is
re-rendered so it doesn't jump visually on the screen.