fix: #95 Re-render on iOS after reordering (#97)
This fixes a regression from #85 which introduced badly-rendered lists
in iOS on subsequent drags. #85, funny enough, was trying to fix #76.

The core cause of this was that `inReorderingRef` was inconsistent
between renders right after reordering, triggered by `reset()`. We now
update the ref prior to the reset. Issue would only reproduce on device,
and only in more complicated apps than `/example` (I suspect because of
re-rendering speeds).

Also figured out we don't need to change keys on the items as long as we
made this change. All in all, seems to work better.

Tested on iOS device by integrating into a more complicated app, and on
Android simulator.