fix: web: manually measure cells when index is updated (#102)
In React Native Web, onLayout is not fired when items are moved around in a FlatList. This causes odd behaviour similar to what has previously been documented in issues #27 #38.

Seemingly this is an underlying bug with RN Web (necolas/react-native-web#2481 - this issue mentions this library specifically).

This PR manually measures the layout of components only on web when the index is updated.